requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.8.2
httpx==0.27.2
openai==1.48.0
tenacity==8.5.0